requirements.txt
GitPython
fastmcp
pydantic